% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethComp.R
\name{check.trans}
\alias{check.trans}
\title{Functions to handle transformations of measurement results.}
\usage{
check.trans(trans, y, trans.tol = 1e-05)
}
\arguments{
\item{trans}{A list of two functions, each other's inverse.}

\item{y}{Vector of numerical values where the functions should be each
other's inverse.}

\item{trans.tol}{Numerical constant indication how precise the evaulation
should be.}
}
\value{
\code{check.trans} returns nothing.
}
\description{
Check whether two functions actually are each others inverse.
}
\author{
Bendix Carstensen, Steno Diabetes Center,
\url{https://bendixcarstensen.com/}.
}
