% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deming.R
\name{Deming}
\alias{Deming}
\title{Perform Deming regression}
\usage{
Deming(
  x,
  y,
  vr = sdr^2,
  sdr = sqrt(vr),
  boot = FALSE,
  keep.boot = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{alpha}{significance level}
}
\description{
Perform Deming regression
}
