% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.measures.R
\name{middle}
\alias{middle}
\title{Function to identify the middle of a vector}
\usage{
middle(w, rm = 1/3)
}
\arguments{
\item{w}{A numeric vector of values}

\item{rm}{A value between 0 and 1 giving the percentage of extreme observations to remove}
}
\value{
A logical vector of indices that a
}
\description{
Function to identify the middle of a vector
}
