% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.MCmcmc.R
\name{pairs.MCmcmc}
\alias{pairs.MCmcmc}
\title{Create a pairs plot for an MCmcmc object}
\usage{
\method{pairs}{MCmcmc}(
  x,
  what = "sd",
  subset = NULL,
  col = NULL,
  pch = 16,
  cex = 0.2,
  scales = "free",
  ...
)
}
\description{
Create a pairs plot for an MCmcmc object
}
