% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbreg.R
\name{predict.PBreg}
\alias{predict.PBreg}
\title{Predict results from PBreg object}
\usage{
\method{predict}{PBreg}(
  object,
  newdata = object$model$x,
  interval = "confidence",
  level = 0.95,
  ...
)
}
\description{
Predict results from PBreg object
}
