% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MCmcmc.R
\name{print.MCmcmc}
\alias{print.MCmcmc}
\title{Print a MCmcmc object}
\usage{
\method{print}{MCmcmc}(x, digits = 3, alpha = 0.05, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{Number of digits to print}

\item{alpha}{Significance level}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a MCmcmc object
}
