% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.MCmcmc.R
\name{subset.MCmcmc}
\alias{subset.MCmcmc}
\title{Subset an MCmcmc object}
\usage{
\method{subset}{MCmcmc}(x, subset = NULL, allow.repl = FALSE, chains = NULL, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{subset}{Numerical, character or list giving the variables to keep.
If numerical, the variables in the \code{MCmcmc} object with
these numbers are selected. If character, each element of the
character vector is "grep"ed against the variable names, and
the matches are selected to the subset. If a list each element
is used in turn, numerical and character elements can be mixed.}

\item{allow.repl}{Logical. Should duplicate columns be allowed in the result?}

\item{chains}{Numerical vector giving the number of the chains to keep.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Subset an MCmcmc object
}
