% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Meth.R
\name{summary.Meth}
\alias{summary.Meth}
\title{Summary for Meth obect}
\usage{
\method{summary}{Meth}(object, ...)
}
\arguments{
\item{object}{A \code{Meth} object.}

\item{...}{Parameters passed on to both the panel function plotting methods against each other, as well as to those plotting differences against means.

rdname summary}
}
\description{
Summary for Meth obect
}
