% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MCmcmc.R
\name{summary.MCmcmc}
\alias{summary.MCmcmc}
\title{Summary}
\usage{
\method{summary}{MCmcmc}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{An \code{MCmcmc} object}

\item{alpha}{1 minus the the confidence level}

\item{...}{Not used}
}
\description{
Summary
}
