% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.measures.R
\name{corr.measures}
\alias{corr.measures}
\title{Classical association measures}
\usage{
corr.measures(x, y)
}
\arguments{
\item{x}{A vector of numeric values of length N}

\item{y}{A vector of numeric values of length N}
}
\value{
A vector of four association measures
}
\description{
A function that returns the values of some of the classical
association measures proposed in the literature
}
